separate perfect_hash_map_builder_mpi
