+++ src/common/kmer_index/extension_index/kmer_extension_index_builder_mpi.hpp
@@ -13,7 +13,7 @@
 #include "projects/hpcspades/common/kmer_index/kmer_mph/kmer_index_builder_mpi.hpp"
 #include "kmer_index/kmer_mph/kmer_splitters.hpp"
 #include "kmer_index/kmer_counting.hpp"
-#include "kmer_index/ph_map/perfect_hash_map_builder.hpp"
+#include "projects/hpcspades/common/kmer_index/ph_map/perfect_hash_map_builder_mpi.hpp"
 #include "io/reads/multifile_reader.hpp"
 
 #include "pipeline/partask_mpi.hpp"
